spotify-consumer/requirements.txt
requests==2.21.0
kafka-python==1.4.5
cassandra-driver==3.17.0
